fw-gate: hoist gold-FW recovery to run-unit-tests harness so it runs before each enable_only (initial + retries); test-fw-update.py reverted to its pre-#15141 simplicity
